Split math from geometry
